Added gate to shadow test only VHD / VPC formats
Utilized the imagefile insepctor to obtain the file format to determine
if the shadowtest path should be used. We only match to vpc since
qemu-img will only output vpc file format despite vhd and vpc being
different extensions for the same format.

Also passing the metadata from the imagefile inspector so that when the
API Inflater is created it can utilize the metadata and not rerun the
insepctor.